Fix issue not being able to retrieve LCCNs starting with letters
ID requires that the LCCN matches what is in the `index:lccn` value.
These are given as `sn 85066387`. If you take out the space, there's no
longer anything to match on. The letters aren't needed because it will
still match on the numbers.

This update checks if the lccn begins with 2 letters. if it does, it
will make sure that there is a space between the letters and the
numbers.