Fix arrows & expand geo search
This reverts the subject builder back to having a separate section for
looking up geo subdivisions. Including it with the topical subject
headings kept surfacing cases where it wouldn't work or wouldn't work as
intended with the fixes for those not being easy, or maybe even possible
without a massive re-work.

This also expands the geo subdivision search to use the more generic
`...id.loc.gov/suggest2...` endpoint rather than only searching on
subjects. This allows the search to find both subjects and names when
building.

This also includes a fix for the arrows not working.