Capture the `GACs` and `marcKey` when possible
This has `extractContextData` in `utils_network.js` to capture the
marcKey.

And updates `setValueComplex` in `profile.js` to load the values GACs
and marcKey into the `userValue` of the profile.

There's a small update to `createLiteral` in `utils_export.js` to be
able to handle the incoming uri for the GACs code. The bf2marc
conversion wants `rdf:about`, but `createLiteral` was only able to
create `rdf:resource`.