Fix some errors that seemed unrelated
The errors were caused by `if` statements that depended on
searchResults, but if we loaded an existing search there wouldn't be any
search results so the error. The error also seems to have made the
`$refs` unreachable causing the hint boxes to not render correctly.

I want to jump out a window.