Fix issue when selecting a new HierGeo
There was some weirdness if someone tried to edit an existing subject
with HierGeo term in it. The searchString had normal hyphens which
caused the selection of a new term to have some pieces of the old term.

For example, Trying to edit "Portugal--Porto" to "Portugal--Porto
(District)" would create "Portugal--Porto (District)--Porto"