Improve MARC output when building with complex sub
Marva maintains the compound subjects created with the builder and
creates XML with them. So building "Korea--History" ends up with the
whole term in the XML.

If it's the only term, that is not an issue for the bf2m conversion. It
will split the term up as appropriate. However, if you add another term
to it, `Korea--History--Fiction`, the conversion leaves `Korea--History`
as one term.

This update will break apart the complex subject into it's pieces when
the "Add" button is pressed.

Still needs work to get the URIs into the XML.