Abort search if there is one active.
This should help address the timing issue with the type ahead. In some
cases, the search would trigger before someone finished typing. This
would result in both searches being performed, but they would be
returned backwards with the results from the earlier search overriding
the most recent.

This is only working in `complex` lookups. It breaks the `subjectSearch`
because that always does multiple searches at the same time; none of
which should be aborted.