Fix bugs with `IntendedAudience`
1) The `fullType` returned by `returnContext()` was `bf:Authority` for
intended audience. This was because that is the `rdf:type` in the rdf
for the record. The fix relies on the `suggectTypeProfile()` function
just like it uses for literasl.
2) There was a weird bug where when you add an additional `IntendedAudience` field, it
would pull the defaults from the `Context` template. I added a check
that uses the URIs of the components to make sure that defaults don't
bleed into other components.