Improve button hiding
There was only a check on that component, but sometimes the component is
part of a group. Because the default insertion function works on the
whole group, we'll do a check to see if any piece of the group supports
defaults to allow the user to add all the defaults of the group at once.

For top level components, we'll only check that selected component has
defaults.