Fix org creator not PrimaryContribution
If an Organization is the primary contributor, it would only be listed
as a contributor at first. Doing it a second time would set it as
primary. The cause was that the active profile would have a type, which
prevented `setTypesForBlankNode()` from using the suggested type. Now if
the suggested type is `PrimaryContribution` it will override the
existing type.