Fix contributor emptying out when switching type
This seems like a naive way of doing things.

The issue was that the carry-over data was being built and stored,
but when the time came to insert it into the new component there was a
failure to match on the properties.

The creator it was trying to match on `sameAs` because this is the
property for the `nextRef` but it saves the carry-over value on the
`bf:agent` property.

The fix here is to add the `propertyPath` uris to the
`possibleProperties` to ensure that they are available for the match.