Fix false positive when match built subject with complex
We try to match a build subject to a complex subject if the person
doesn't pick the complex one for some reason. However, there are cases
where the strings can match but the types won't. This will now check the
string and then the types to make sure that there is a match on both.