Improve bibId in title
Moved logic to a function.
`mounted` is needed to populate the title when loading from a search.
However, relying only on this can cause the wrong bibID to be loaded
when loading from the records list. The `activeProfile` when loading
from the record list will be the last record opened.

`updated` updates the bibID in the title to reflect the currently open
record. Not sure why this isn't enough locally.