Fix cutter tool freezing
Still not sure what the issue was exactly, but adding a bunch of subject components and then deleting some of them could cause the freezing.

There were two console errors in `ComplexLookupModal.vue` -> `selectChange()`. `toLoad` was null and causing issues in the function.

They have been fixed by adding logic to handle a null value in `toLoad` and the cutter issue seems resolved.