Fix navigation when there are mulitple instances
When there are mulitple instances, they were the same except for the
profile name. But that meant that all of their components, looked the
same which interfered with using the property panel to navigate.

This makes the `parent` & `parentId` property of each component match
the profile name so they are unique.

This is based on the code for generating additional secondary instances.