Fix subject editor now allowing changes
The issue was happening when a new piece of the subject would contain a
letter that was in some part of the subject that came before in a
complex subject. For example,

`Recreation--Serbia--Belgrade--History--20th century` when trying to add
`--exhibition` to the end, there would be a clash because `e` is in the
complex geo component, and when stitching things together the new `e`
would get shoved into the existing component and disappear from the
string.