Create additional field from empty
If you try to create "Additional Literal" on an empty field, you need to
run the function a second time inorder for it to create the new field.
The first time will populate the record with "new value" and the second
will create a new field and give it "new value".