Tune Classifiction for defaults
There were some issues related to Classification and defaults. Entering
them for an LCC, would result in the defaults for all of the different
Class schemes being parsed and entered incorrectly.

There was also an issue with the DDC edition number not populatin
correctly. The field would be empty and the XML would be incorrect.