Improve working with incoming GEOs in builder
If a record comes in with a heading like
`topic`--`Country`--`City`--`topic`, where the Country and City aren't
connected, the builder was causing more work than it should to allow a
cataloger to put the Country and City together. Switching to `Indirect
Geo` and retyping the `--` between the Country and City would create a
single `Country--City` element, but it would also make any terms
following this unverified. Now that verification should stick.