Update `setValueComplex()` & `setComplexValue()` to work with sugg2
This is trying to avoid a call to `returnContext()` in
`setValueComplex()` since we are getting the types from the suggest2
service now. We get a list of types from the suggest2 service, and use
the first one.

It will work as it has for cases when there is no type passed. Such as
with `Hub` creation.