Adjust how XMl for relation/Series is created
`Input Transcriped Series` would be labeled as `bflc:Uncontrolled` with
no indication that it is a series.

In `utils_parse.js/transformRts()`, there's an expection for `bflc:Uncontrolled` to build the element with the `tagName` of the
element instead of with the value of`rdf:resource`. This creates `<bf:Series>` where it should.
This triggers when the element is `bf:relation` && value of
`rdf:resource` would be `bflc:Uncontrolled`.

In `utils_export.js/buildXMLProcess()`, theres an exception when
`bnodeLvl1.tagName == bf:Relation && bnodeLvl2.tagName == bf:Series &&
pLvl2.tagName == bf:associatedResource` to add an `rdf:type
=bflc/Uncontroled`.