Make order more universal
Before it would only really work with records that have the same
components. So if you saved an order that only had 1 subject, and then
loaded a record wit >1 subject, only 1 subject would show up.

This also keeps similar elements together. But the `primary` component
is that one that needs to be moved. `primary` meaning the component with
an unedited component id. `id` not `id_<date>`