Fix complex headings not appearing
If a record is loaded with a complex heading and the user adds
subdivisions to it, clicking on the complex heading would fail to show
the authorized form in the results. This was because of how the search
was conducted. The complex search used the whole string, original
headding + subdivision, which would not return the original heading by
itself.

This change adds a search to look for complex headings, terms that have
`--` in them, in addition to the other complex lookup.