Fix `Jurisdictions` being of type `Organization`
When a subject was a jurisdiction, it was being given the type
`Organization` which was throwing off the conversion. This was because
`Jurisdiction` is defined in the `collection` but type was determined
solely by the values in `rdfTypes`.

Now, we'll do a backup check of the `collections` and set the type to
`Jurisdiction` when appropriate.