Adjust how Marva handles hitting the end of the results
The pagination depends on the number of `total` results that `suggest2`
says it found. But sometimes this number can be misleading. The number
is based on the number of results prior to deduping. The result can be
that Marva says there are `5` pages, but there are really only `3`. When

If Marva doesn't have any hits, the search could become unusable because
the offset/page get set too far, but also the page navigation will
disappear. Instead, if there are no results and the offset is > 0, the
user will be sent back to the previous page.