Fix `bf:duration`
If the duration was given an `datatype URI` in the profile, entering a
value in Marva would cause a crash. This was because the URI would be
placed into `@type` and then Marva would understand the `userValue` as a
bNode in `utils_export.js`. When it tried to `createLiteral()` it would
run into an issue and crash. Additionally, for `bf:duration` it needs an
`@datatype` to generate the correct XML, but nothing in Marva tries to
make an `@datatype`

The fix is an exception in `setTypesForBlankNode()` that will set the
`@datatype` instead of `@type`. This fixes the crash and generates the
correct XML. "Correct" meaning it follows the pattern for `bf:duration`
in the ontology example.