Update how `indirect geo` is input
This tries to account for more usecases. Namely when the cataloger
selects the first part of an indirect geo from the LCNAF/LCSH input then
swaps to indirect geo to finish.

1) The cataloger enters: `Topic--Country--City` and validates Topic and
Country then jumps to `Indirect Geo`. The search string will be changed
so that the `Country--City` are one entry and run the search on that.

2) Cataloger enters: `Topic--Country` and validates each then swaps to indirect geo and
types `--City`. The dashes before `City` will be swapped for the "not
dashes" that indirect geo uses.