Add check for `type` in `<Ref>`
Perform a check against `rtLookup` with the `useId` as the Key and
compare the `resourceUri` to `userValue[@type]`. If there is no match,
we'll take it to mean that the `rdf:type` for the component isn't set to
an available option. We'll adjust the class of that component so that it
is highlighted in red.

The error will also get picked up by the Validation and incorporated
into the results with the ability to jump directly to that component.