Adjustment to parsing hub subjects and XML creation
Parsing: The Hub subjects was being overwritten by the `rdf:about` for the hub.
There's an exception that will prevent this.

XML: There appears to have been a typo(?) in the creation of the `pLvl4`
that was using `key2` instead of `key3` this was causing some weird XML
`bf:contribution` > `bf:Contribution` > `bf:contribution` when the last
piece should have been `bf:agent`