Some rework to how existin record check works
The LCCN check works the same: it uses the value in the LCCN input
field. But now there's an option for "Other Identifier." When selected
it'll populate the field with the value of the search query. But it can
be adjusted to check something else.

This does 2 things:
1) It makes explicit something that was happening under the hood.
2) Allows the cataloger to make adjustments incase there's something
else they would like to search on.

Populating it with the `wcQuery` is a convience so they have something
they are starting with.