Comment out code for setting components to load with a record
There's a lot that could go wrong here. The current implementation could
duplicate information because it is additive and doesn't do any
checking of existing data.

There's also the issue of what happens when a record is reopened in
Marva. If components are set to load with the record, they'll load each
time the record is loaded.