NAR from action button works like from complex modal
If there's a name present and someone uses the "Create NAR" option from
the action button it will function more like the the "Create NAR" option
available in the complex lookup modal. The field in the NAR creation
form will populate with name and the paste buttons will show up and have
that name as their value too.