Marva is not giving a failed to save message
At least in dev, when there is no backend to save to Marva just
continues to wait for a response from the server that it never gets and
doesn't deliver a "failed to save alert." This adds a 3 second timeout
to the save. There's also some changes to how the icon changes to
reflect whether or not something saved.