Limit to `PZ`
I think it's only PZ that can include title in the number.

In subclass PZ, works by individual authors are subarranged by work letters
representing the title of the work, ignoring initial articles. These title work letters are not Cutter
numbers, because the capital letter is not followed by a numeral. Instead, the first letters of the
title of the work (excluding initial articles) are used. -
Classifciation and Shelflisting Manual G350