Fix literal names remaining literal
If a record came in with a literal name as a subject, when trying to
replace it with the Authority form the literal one could remain. This
could happen if the literal and the authority form of the name is the
same. Now it should prefer the authority form over the literal